Handle missing `service_node_privkey` for SN endpoint
`at` throws an exception if the key does not exist. When an Oxen node is
not running in SN mode, the key is missing from the output so using `at` throws.

There's a helpful error message in the output that hints that perhaps
the node is not running in SN mode that is no longer reachable in this
case. We can make it reachable by avoiding `at` and querying the key
using `find`.