Fix initial rpc.get_block_hash failing on Oxen 11
The bootstrapping rpc.get_block_hash on startup where it queries the
latest hashes for storage tests was not working on Oxen 11 because the
RPC has changed to use a 'heights' instead of 'height' array. In Oxen 10
'height' was used but not anymore. The endpoint also now returns the
hash under the the height as a key in the response object whereas
previously the endpoint returned the hex hash directly without any JSON.

The number of general threads assigned to OMQ has been explicitly
declared and documented. Upon reviewing the code I was under the
impression that there were several race conditions between these async
OMQ requests we were sending, however, it turns out when I tried to fix
it, first by making the OMQ request block, the storage server would
deadlock. Cue me finding out that we only allocate 1 worker thread to OMQ.

This means that some of the storage server code has been written in a way that
assumes that it can access objects across threads without the necessary
synchronisation. So I've marked it as such that changing that would potentially
introduce a lot of bugs and that the current instances of SS violating
it are safe even though it may not look like it from the outset.