Fix infinite loop on failed initial update_swarms
If the initial `update_swarms` call on startup in `on_oxend_connected`
fails because of a network or parsing error then the infinite loop that
waits for the `first_response_cv` would never get triggered and
infinitely loop waiting for the condition variable to be notified.

I've simplified this 'async-await' pattern using promises and futures
which is another pattern we also use in other parts of this codebase.
This removes the need for `first_response_cv` and `first_response_mutex`.

If the initial update errors, this can be detected and the update can
then initial swarm update can be reattempted.