Storage server contact & initial data refactor
This PR refactors how storage server tracks the contact information (ip,
port, pubkeys) of network nodes and how it initiates contacts with other
nodes on the network.

In particular, this aims to make new swarm joining more reliable:
currently there are known issues where new swarm members are not
receiving initial swarm messages because the other swarm members can
send the data to the new node before the new node is ready to recognize
them as members.

This reworks the mechanism:

When a new swarm member is detected, existing swarms members now send
periodic (new) "snode.data_ready" requests to the new member(s), and do
not attempt to push swarm data to that member until they get back a
successful reply (thus indicating that we can reach the node, and that
the node properly recognizes us as a swarm peer).  Previously members
would attempt to push pretty much right away, but there are various ways
the new node might either be unreachable, or might reject messages from
a node it doesn't recognize.

If the command times out, or the remote returns anything other than
"OK", the new member is returned to the queue and contact will be
reattempted after another ~30s.

Once the OK is received, existing swarm members then push swarm messages
to the new member via the existing snode.data endpoint for the new
member to load the data.

(This is not a perfect solution, and there are still various issues
remaining, but this is a stop-gap while we also develop a more robust,
active syncing mechanism).

Along the way this ends up touching a lot of other code:

- the "Swarms" class had a very confused triple purpose of tracking the
  current swarm, tracking all snodes, and tracking contact info, and was
  confusing to use.  These three purposes are now in three separate
  (though related) classes:
  - Swarm tracks the current swarm info, i.e. current swarm members.
  - Network tracks the global service node state, i.e. can look up which
    swarm a user pubkey or service node pubkey belongs to.
  - Contacts tracks the contact info of all registered service nodes,
    regardless of their state (aside from being registered).
  - sn_record is gone, and it *somewhat* replaced with a contact.  This
    isn't a direct replacement, though: many places that previously had
    a sn_record now simply take a legacy_pubkey, and look up the contact
    when needed via the Contacts object rather than passing the contact
    around.

- Storage server now subscribes to address updates from oxend, so that
  it can update the contact info for a node immediately when a proof
  updates it, rather than only doing it when new blocks come in (which
  can be 2 minutes old during normal times, and much longer if there is
  a chain delay).

- We now fetch all SNs from oxend (previously only fully-funded ones):
  this lets us set up contacts and permissions for partially funded
  nodes, even though they aren't yet in any swarm, but then we are more
  likely to have them ready to go as soon as they get fully funded.

- Bump oxen-mq required version to 1.2.19, for an update that rechecks
  SN state on each command, rather than only when establishing a new
  connection.

- Replace all remaining std::stringstream usage with fmt.

- Changed the legacy_keypair (and similar for ed, x) from a std::pair
  typedef to a standalone struct with `.pub` and `.sec` members.  The
  plain pair was a bit dangerous because it wasn't immediately obvious
  whether .first was the pubkey or the privkey.  This also pushes it
  into some places that were still taking them as two separate
  arguments.
  - Related change: we now initialize from oxend with keypairs instead
    of just secret keys as we just had to compute all the pubkeys
    elsewhere anyway, so easier to just compute the pubkeys as soon as
    we get the secret keys.

- Added `load_from_bytes`/`load_from_hex` member functions to key types,
  and added `[[nodiscard]]` to the static `from_bytes`/`from_hex`.
  Previously it was easy to accidentally use `instance.from_hex(...)`
  which just calls the static method and doesn't actually change
  `instance`.

- Removed some string helper functions that are trivial now that we have
  fmt, added `get_human_readable_bytes` from oxen-core.

- Made the "random tests per tick" into a variable so that it can be
  changed easily later; previously the code was only ever trying 1.

- Replaced bt_list serialization with a more efficient bt_list_producer,
  along with a hack that lets us extract the final version-byte-prefixed
  string without needing extra allocations/copies.

- Remove the exclusive transaction from the save_bulk call (which was
  locked multiple times over): there doesn't seem to be any reason why
  this needs to be exclusive, blocking out other endpoints while the
  data gets loaded (which might be length as we could have 9MB of data
  to insert).

- Add some code documentation to better describe why we do what we do
  during swarm dissolution and new network swarms.