Add SS version to contact; gate pre-transfer handshake on version
This adds retrieval and tracking of the remote storage server version to
the `contact` class so that we can use it when determining how we
interact with a peer.

This commit then uses that to gate our support for a pre-transfer
handshake on the remote running the new storage server version.

Also:
- bumps version to 2.10.0
- switches how we request service node info from the local oxend to use
  the simpler array-of-fields rather than the oxen-10 and earlier
  dict-of-true-values.