Match comment w/ code regarding when to add retryable request on failure
Comment states that only timed-out requests are retried. This is correct
as an error response with error code and text are stiuations that might
mean that the recipient node is not in a valid state or will ever accept
the request in which case the safe default is to not retry to that node.

It is possible in future since all possible error states are known to
handle them specifically for the command. But for now, a sane default is
to only allow retries to nodes that were offline or we failed to
communicate with.